yuzu: Rename check_vulkan to startup_checks
