+++ src/yuzu/startup_checks.cpp
