+++ src/yuzu/startup_checks.h
